# docs/swagger.yaml
basePath: /
definitions:
  auth.Login:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.LoginBody:
    properties:
      payload:
        $ref: '#/definitions/auth.Login'
    type: object
  auth.LoginResponse:
    properties:
      accStatus:
        type: integer
      accessToken:
        type: string
      error:
        type: string
      refreshToken:
        type: string
      role:
        type: string
      userId:
        type: string
    type: object
host: droidzed.tn
info:
  contact:
    email: droid.zed77@outlook.com
    name: Aymen Dhahri
    url: https://droidzed.tn/support
  description: This is the GoLance API documentation.
  license:
    name: MIT
    url: https://github.com/DroidZed/my_blog/LICENSE
  termsOfService: https://droidzed.tn/terms/
  title: My Website's API Docs
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Get token, user basic data
      parameters:
      - description: Login User
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth.LoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth.LoginResponse'
      summary: Auth user
      tags:
      - auth
  /api/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh access + refresh tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth.LoginResponse'
      security:
      - Bearer: []
      summary: Refresh tokens
      tags:
      - auth
swagger: "2.0"